+++ com/planet_ink/coffee_mud/WebMacros/RoomData.java
@@ -849,7 +849,7 @@ public class RoomData extends StdWebMacro
 			@Override public List<MultiPartData> getMultiParts() { return httpReq.getMultiParts(); }
 			@Override public double getSpecialEncodingAcceptability(String type) { return httpReq.getSpecialEncodingAcceptability(type); }
 			@Override public String getFullHost() { return httpReq.getFullHost(); }
-			@Override public List<int[]> getRangeAZ() { return httpReq.getRangeAZ(); }
+			@Override public List<long[]> getRangeAZ() { return httpReq.getRangeAZ(); }
 			@Override public void addFakeUrlParameter(String name, String value) { params.put(name.toUpperCase(), value); }
 			@Override public void removeUrlParameter(String name) { params.remove(name.toUpperCase()); }
 			@Override public Map<String,Object> getRequestObjects() { return httpReq.getRequestObjects(); }
